Compatible with 32-bit

+++ internal/util/util.go
@@ -69,8 +69,8 @@ func UintToU16(u uint) (uint16, error) {
 }
 
 func UintToU32(u uint) (uint32, error) {
-	if u > math.MaxUint32 {
-		return 0, fmt.Errorf("u %d is greater than %d", u, math.MaxUint32)
+	if int64(u) > int64(math.MaxUint32) {
+		return 0, fmt.Errorf("u %d is greater than %d", u, int64(math.MaxUint32))
 	}
 	val := uint32(u)
 	return val, nil
@@ -104,8 +104,8 @@ func IntToU16(u int) (uint16, error) {
 }
 
 func IntToU32(u int) (uint32, error) {
-	if u > math.MaxUint32 {
-		return 0, fmt.Errorf("u %d is greater than %d", u, math.MaxUint32)
+	if int64(u) > int64(math.MaxUint32) {
+		return 0, fmt.Errorf("u %d is greater than %d", u, int64(math.MaxUint32))
 	} else if u < 0 {
 		return 0, fmt.Errorf("u %d is less than 0", u)
 	}
